move validators folder to models folder
